Crate comlib_io

Comlib Input/Output Utilities

This library contains commonly needed utilities for handling IO.

The most commonly needed API is the Input struct which can be used to parse input. The following code reads a line count n from the standard input and then reads the following n lines, each containing a pair of numbers.

use comlib_io::*;
let stdin = std::io::stdin();
let mut input = Input::from_stdin(&stdin);
let n = input.match_line(input_pattern!(usize));
let queries = input.match_lines(input_pattern!(usize, " ", usize), n..=n);

The library also contains some useful utilities for printing out multiple values. Currently the following wrappers are supported:

Writing space-separated values

Macros

input_pattern
Pattern for parsing input

Structs

Input
Helper for reading objects implementing InputPattern trait.

Traits

Consumable
Value which can be greedily consumed from &str.

InputPattern
Parser pattern for parsing input.

Functions

spaced
Wrapper for displaying items of iterator as separated by spaces.