Function comlib_io::spaced

pub fn spaced<I>(i: I) -> SpacedWriter<<I as IntoIterator>::IntoIter> where
    I: IntoIterator,
    <I as IntoIterator>::IntoIter: Clone,
    <I as IntoIterator>::Item: Display, 

Wrapper for displaying items of iterator as separated by spaces.

Note that this does not mean that the Debug would format the values as separated by spaces.

Examples

assert_eq!(format!("{}", spaced(vec![1, 2, 3])), "1 2 3");