Macro comlib_io::input_pattern

macro_rules! input_pattern {
    () => { ... };
    ($($pattern : tt) +) => { ... };
}

Pattern for parsing input

The patterns matched greedily from the start. Each typed input is matched for as long prefix as possible. If the typed input is followed by a string pattern, then the type is matched until the first occurrence of that pattern, otherwise the type is matched greedily for as long prefix of the input as possible. For numeric types this means until the first non-numeral character, and for strings until the end of the string. Note that for numeric types this doesn’t take into account that the value fits into the given type.

The pattern may contain any type which implements FromStr when it’s followed by a string pattern, or any type which implements Consumable when it’s not followed by a string pattern.

The parser can parse variable number of occurrences of the pattern as Vecs. The variable arguments are enclosed in [brackets] and can contain any valid pattern, including more vectors.

A string pattern can be made optional, in which case it is not necessary that it occurs in the input. Greedy matching is still stopped at the first occurrence of the pattern if it exists. Optional patterns are especially useful with vector patterns and to match plurals of words.

See examples on how to use the input_pattern.

Examples

use comlib_io::{input_pattern, InputPattern};

// Parse two numerals separated by a space
assert_eq!(input_pattern!(usize, " ", usize).parse_all("1 2"), Some((1, 2)));

// Parse key-value pair
assert_eq!(input_pattern!(String, ": ", usize).parse_all("key: 2"), Some(("key".to_string(), 2)));

// Parse a vector of space-separated numbers
assert_eq!(input_pattern!([usize, " "?]).parse_all("1 2 3 4"), Some(vec![1, 2, 3, 4]));

// Parse a vector of key-value pairs
assert_eq!(
    input_pattern!([String, " = ", String, " "?]).parse_all("key1 = v1 key2 = v2"),
    Some(vec![
        ("key1".to_string(), "v1".to_string()),
        ("key2".to_string(), "v2".to_string()),
    ])
);

// Without the optional space separator in the pattern, the string won't be matched
assert_eq!(
    input_pattern!([String, " = ", String, " "]).parse_all("key1 = v1 key2 = v2"),
    None
);

// Match both singular and plural form
assert_eq!(
    input_pattern!([usize, " ", String, " item", "s"?, ", "?]).parse_all("1 red item, 2 blue items"),
    Some(vec![
        (1, "red".to_string()),
        (2, "blue".to_string()),
    ])
);