Trait comlib_io::Consumable

pub trait Consumable where
    Self: Sized, {
    type InputError;
    fn consume(input: &str) -> Result<(Self, &str), Self::InputError>;
}

Value which can be greedily consumed from &str.

Associated Types

type InputError

Type for representing errors which can happen during consuming values.

Loading content...
Required methods

fn consume(input: &str) -> Result<(Self, &str), Self::InputError>

Consume input from the given &str.

In case of success, returns the consumed value and the rest of the string. In case of an error, returns the error.

Loading content...